requirements.txt
openai==1.96.1
langchain-core==0.3.69
langchain-openai==0.3.28
langgraph==0.5.3
langchain-mcp-adapters==0.1.9
typing-extensions==4.14.1
python-dotenv==1.1.1
openai-agents==0.2.0
PyYAML==6.0.1
fastapi>=0.115.8
uvicorn>=0.34.0